Vaina

+++ algoritmos.hh
@@ -23,7 +23,7 @@ Nodo* a_static555(int8 estado[16], int peso);
 std::list<Nodo*> ida_manhattan(int8 estado[16]);
 
 // Heurísticas
-int heuristica(Nodo* n);
+int8 heuristica(Nodo* n);
 int h_static555(int8 estado[16]);
 int h_static55_M(int8 estado[16]);
 int h_static5_M(int8 estado[16]);
